requirements.txt
appnope==0.1.2
asgiref==3.4.1
attrs==20.3.0
autopep8==1.6.0
backcall==0.2.0
bleach==3.3.1
boto3==1.20.10
botocore==1.23.10
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
decorator==5.0.7
dj-database-url==0.5.0
Django==3.2.9
django-appconf==1.0.5
django-bleach==0.7.0
django-cors-headers==3.10.1
django-crispy-forms==1.12.0
django-environ==0.4.5
django-filter==21.1
django-js-asset==1.2.2
django-messages==0.6.0
django-mptt==0.12.0
django-mptt-admin==2.1.0
django-multiselectfield==0.1.12
django-on-heroku==1.1.2
django-otp==1.1.1
django-pipeline==2.0.7
django-polymorphic==3.0.0
django-polymorphic-tree==2.0
django-postman==4.2
django-ratelimit==3.0.1
django-resized==0.3.11
django-sass-processor==1.1
django-ses==2.3.1
django-storages==1.12.3
django-tag-parser==3.2
django-tinymce==3.3.0
django-widget-tweaks==1.4.8
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
future==0.18.2
gunicorn==20.1.0
idna==2.10
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
jellyfish==0.8.2
Jinja2==3.0.3
jmespath==0.10.0
lib50==2.0.9
libsass==0.21.0
markdown2==2.4.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
packaging==21.0
parso==0.8.2
paypalrestsdk==1.13.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.18
psycopg2==2.9.3
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pycodestyle==2.8.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.5
python-dotenv==0.19.2
python-http-client==3.3.4
pytz==2021.3
PyYAML==5.4.1
qrcode==7.3
rcssmin==1.0.6
requests==2.25.1
rjsmin==1.1.0
s3transfer==0.5.0
sendgrid==6.9.3
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
submit50==3.0.3
termcolor==1.1.0
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.3.0
